Copy the failing job's steps, not a link to them
- Copying failing CI pasted the check rows and nothing else: it
  asked for the whole run's log, and `gh` refuses that while any
  job in the run is still going, which is when it is asked
- A failing check names its own job, so that job is what is read
  now, and a run of fifty jobs where one failed no longer offers
  the other forty-nine
- A run still in progress falls back to the API, which serves a
  finished job's log whatever the rest of the run is doing
